# internal/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: lowcode-database API
  description: |
    Postgres-backed low-code table service: dynamic tables, columns, rows, indexes,
    relations, and saved queries.

    Optional HTTP webhooks: `/v1/admin/webhooks`. Out of scope (schema bundle import,
    plugins, graph expand, ENUM, RBAC): see docs/roadmap.md.

    **Multi-tenant:** send `X-Tenant-Id` on every request when `TENANT_MODE=multi`.
    In single-tenant mode the header is optional (default tenant from env).

    **Playground:** [lowcode-database-playground](https://github.com/monoposer/lowcode-database-playground) — AG Grid UI (separate repo).

    Errors are returned as plain text (`http.Error`), not JSON envelopes.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local dev (see HTTP_ADDR in .env)
...
paths:
...
  /v1/admin/tables:
    get:
      tags: [Tables]
      summary: List tables
      operationId: listTables
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTablesResponse"
        "400":
          $ref: "#/components/responses/Error"
    post:
      tags: [Tables]
      summary: Create table
      operationId: createTable
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTableRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTableResponse"
        "400":
          $ref: "#/components/responses/Error"
  /v1/admin/tables/{id}:
    delete:
      tags: [Tables]
      summary: Delete table
      operationId: deleteTable
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteTableResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/admin/tables/{id}:rename:
    post:
      tags: [Tables]
      summary: Rename table
      operationId: renameTable
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameTableBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenameTableResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/admin/tables/{tableName}/schema:
    get:
      tags: [Tables]
      summary: Get table schema (columns + indexes)
      operationId: getTableSchema
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/TableNamePath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetTableSchemaResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /v1/admin/columns:
    get:
      tags: [Columns]
      summary: List columns for a table
      operationId: listColumns
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - name: table_name
          in: query
          required: true
          schema:
            type: string
          description: Logical table id.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListColumnsResponse"
        "400":
          $ref: "#/components/responses/Error"
    post:
      tags: [Columns]
      summary: Add column
      operationId: addColumn
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddColumnRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddColumnResponse"
        "400":
          $ref: "#/components/responses/Error"
  /v1/admin/columns/{id}:
    patch:
      tags: [Columns]
      summary: Update column
      operationId: updateColumn
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateColumnBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateColumnResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Columns]
      summary: Delete column
      operationId: deleteColumn
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteColumnResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/admin/indexes:
    get:
      tags: [Indexes]
      summary: List indexes for a table
      operationId: listIndexes
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - name: table_name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListIndexesResponse"
        "400":
          $ref: "#/components/responses/Error"
    post:
      tags: [Indexes]
      summary: Create index
      operationId: createIndex
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIndexRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateIndexResponse"
        "400":
          $ref: "#/components/responses/Error"
  /v1/admin/indexes/{id}:
    get:
      tags: [Indexes]
      summary: Get index by id
      operationId: getIndex
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetIndexResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Indexes]
      summary: Delete index
      operationId: deleteIndex
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteIndexResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /v1/admin/relations:
    get:
      tags: [Relations]
      summary: List relations
      operationId: listRelations
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - name: table_name
          in: query
          schema:
            type: string
          description: Filter by table id (optional).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListRelationsResponse"
        "400":
          $ref: "#/components/responses/Error"
    post:
      tags: [Relations]
      summary: Create relation
      operationId: createRelation
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRelationRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRelationResponse"
        "400":
          $ref: "#/components/responses/Error"
  /v1/admin/relations/{id}:
    delete:
      tags: [Relations]
      summary: Delete relation
      operationId: deleteRelation
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteRelationResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/admin/queries:
    get:
      tags: [Queries]
      summary: List saved queries
      operationId: listQueries
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - name: table_name
          in: query
          schema:
            type: string
          description: Filter by table id (optional).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListQueriesResponse"
        "400":
          $ref: "#/components/responses/Error"
    post:
      tags: [Queries]
      summary: Create a saved query
      operationId: createQuery
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQueryRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateQueryResponse"
        "400":
          $ref: "#/components/responses/Error"
  /v1/admin/queries/{id}:
    get:
      tags: [Queries]
      summary: Get a saved query
      operationId: getQuery
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetQueryResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Queries]
      summary: Update a saved query
      operationId: updateQuery
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQueryBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateQueryResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Queries]
      summary: Delete a saved query
      operationId: deleteQuery
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteQueryResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    TenantIdHeader:
      name: X-Tenant-Id
      in: header
      required: false
      description: |
        Tenant identifier. Required in multi-tenant mode (`TENANT_MODE=multi`) to route
        to the tenant's Postgres database. Also accepted as `X-Tenant-ID`.
      schema:
        type: string
        example: default
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
    TableNamePath:
      name: tableName
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: Error message (plain text from http.Error)
      content:
        text/plain:
          schema:
            type: string
            example: table_name query parameter is required
  schemas:
...
    Table:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        schemaName:
          type: string
        tableName:
          type: string
        idType:
          type: string
          description: Logical type id of the physical row id column (from PG catalog).
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Column:
      type: object
      properties:
        id:
          type: string
        tableName:
          type: string
        name:
          type: string
        typeId:
          type: string
        pgColumn:
          type: string
        isNullable:
          type: boolean
        position:
          type: integer
          format: int32
        config:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Index:
      type: object
      properties:
        id:
          type: string
        tableName:
          type: string
        name:
          type: string
        pgIndex:
          type: string
        columnIds:
          type: array
          items:
            type: string
        isUnique:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateTableRequest:
      type: object
      properties:
        name:
          type: string
        schemaName:
          type: string
        idType:
          type: string
          description: Built-in column type id for the physical row id column (default uuid).
    CreateTableResponse:
      type: object
      properties:
        table:
          $ref: "#/components/schemas/Table"
    DeleteTableResponse:
      type: object
    ListTablesResponse:
      type: object
      properties:
        tables:
          type: array
          items:
            $ref: "#/components/schemas/Table"
    RenameTableBody:
      type: object
      properties:
        newName:
          type: string
      description: Table id is taken from the path (`/v1/admin/tables/{id}:rename`).
    RenameTableResponse:
      type: object
      properties:
        table:
          $ref: "#/components/schemas/Table"
    GetTableSchemaResponse:
      type: object
      properties:
        table:
          $ref: "#/components/schemas/Table"
        columns:
          type: array
          items:
            $ref: "#/components/schemas/Column"
        indexes:
          type: array
          items:
            $ref: "#/components/schemas/Index"
...
    SortOrder:
      type: object
      properties:
        attribute:
          type: string
        sortOrder:
          type: string
          enum: [ASC, DESC]
...
    AddColumnRequest:
      type: object
      required: [tableName]
      properties:
        tableName:
          type: string
        name:
          type: string
        typeId:
          type: string
        isNullable:
          type: boolean
        position:
          type: integer
          format: int32
        config:
          type: object
          additionalProperties: true
    AddColumnResponse:
      type: object
      properties:
        column:
          $ref: "#/components/schemas/Column"
    UpdateColumnBody:
      type: object
      properties:
        name:
          type: string
        isNullable:
          type: boolean
          nullable: true
        position:
          type: integer
          format: int32
        config:
          type: object
          additionalProperties: true
      description: Column id is taken from the path.
    UpdateColumnResponse:
      type: object
      properties:
        column:
          $ref: "#/components/schemas/Column"
    DeleteColumnResponse:
      type: object
    ListColumnsResponse:
      type: object
      properties:
        columns:
          type: array
          items:
            $ref: "#/components/schemas/Column"
    CreateIndexRequest:
      type: object
      properties:
        tableName:
          type: string
        name:
          type: string
        columnIds:
          type: array
          items:
            type: string
        isUnique:
          type: boolean
    CreateIndexResponse:
      type: object
      properties:
        index:
          $ref: "#/components/schemas/Index"
    GetIndexResponse:
      type: object
      properties:
        index:
          $ref: "#/components/schemas/Index"
    DeleteIndexResponse:
      type: object
    ListIndexesResponse:
      type: object
      properties:
        indexes:
          type: array
          items:
            $ref: "#/components/schemas/Index"
    Relation:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        kind:
          type: string
          description: ONE_TO_ONE | ONE_TO_MANY | MANY_TO_ONE
        sourceTableName:
          type: string
        sourceColumnId:
          type: string
        targetTableName:
          type: string
        targetColumnId:
          type: string
        config:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateRelationRequest:
      type: object
      properties:
        name:
          type: string
        kind:
          type: string
        sourceTableName:
          type: string
        sourceColumnId:
          type: string
        targetTableName:
          type: string
        targetColumnId:
          type: string
        config:
          type: object
          additionalProperties: true
    CreateRelationResponse:
      type: object
      properties:
        relation:
          $ref: "#/components/schemas/Relation"
    ListRelationsResponse:
      type: object
      properties:
        relations:
          type: array
          items:
            $ref: "#/components/schemas/Relation"
    DeleteRelationResponse:
      type: object
    Query:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        label:
          type: string
        tableName:
          type: string
        filter:
          type: object
          additionalProperties: true
        sort:
          type: array
          items:
            $ref: "#/components/schemas/SortOrder"
        columnIds:
          type: array
          items:
            type: string
        config:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateQueryRequest:
      type: object
      properties:
        name:
          type: string
        label:
          type: string
        tableName:
          type: string
        filter:
          type: object
          additionalProperties: true
        sort:
          type: array
          items:
            $ref: "#/components/schemas/SortOrder"
        columnIds:
          type: array
          items:
            type: string
        config:
          type: object
          additionalProperties: true
    CreateQueryResponse:
      type: object
      properties:
        query:
          $ref: "#/components/schemas/Query"
    ListQueriesResponse:
      type: object
      properties:
        queries:
          type: array
          items:
            $ref: "#/components/schemas/Query"
    GetQueryResponse:
      type: object
      properties:
        query:
          $ref: "#/components/schemas/Query"
    UpdateQueryBody:
      type: object
      properties:
        label:
          type: string
        filter:
          type: object
          additionalProperties: true
        sort:
          type: array
          items:
            $ref: "#/components/schemas/SortOrder"
        columnIds:
          type: array
          items:
            type: string
        config:
          type: object
          additionalProperties: true
      description: Query name is taken from the path.
    UpdateQueryResponse:
      type: object
      properties:
        query:
          $ref: "#/components/schemas/Query"
    DeleteQueryResponse:
      type: object
...